+++ pkg/consts/aigateway.go
@@ -47,9 +47,9 @@ const (
 	RelatedImageAIGatewayDataPlaneEnvVar = "RELATED_IMAGE_AIGW"
 
 	// DefaultAIGatewayDataPlaneBaseImage is the base image name for the AI Gateway container.
-	DefaultAIGatewayDataPlaneBaseImage = "kong/kong-ai-gateway-dev"
+	DefaultAIGatewayDataPlaneBaseImage = "kong/kong-ai-gateway"
 	// DefaultAIGatewayDataPlaneTag is the default image tag for the AI Gateway container.
-	DefaultAIGatewayDataPlaneTag = "2.0.3-rc.1" // renovate: datasource=docker depName=kong/kong-ai-gateway-dev
+	DefaultAIGatewayDataPlaneTag = "2.0.3" // renovate: datasource=docker depName=kong/kong-ai-gateway
 	// DefaultAIGatewayDataPlaneImage is the full default image reference for the AI Gateway container.
 	DefaultAIGatewayDataPlaneImage = DefaultAIGatewayDataPlaneBaseImage + ":" + DefaultAIGatewayDataPlaneTag
 )